matterdb::access

Trait AccessExt

pub trait AccessExt {
    type Base<'a>: RawAccess
       where Self: 'a;
    type Ref<'a>: Access<Base = Self::Base<'a>>
       where Self: 'a;

    // Required method
    fn get_ref(&self) -> Self::Ref<'_>;

    // Provided methods
    fn get_group<'s, K, I>(
        &'s self,
        name: impl Into<String>,
    ) -> Group<Self::Ref<'s>, K, I>
       where K: BinaryKey + ?Sized,
             I: FromAccess<Self::Ref<'s>> { ... }
    fn get_entry<I, V>(&self, addr: I) -> Entry<Self::Base<'_>, V>
       where I: Into<IndexAddress>,
             V: BinaryValue { ... }
    fn get_list<I, V>(&self, addr: I) -> ListIndex<Self::Base<'_>, V>
       where I: Into<IndexAddress>,
             V: BinaryValue { ... }
    fn get_map<I, K, V>(&self, addr: I) -> MapIndex<Self::Base<'_>, K, V>
       where I: Into<IndexAddress>,
             K: BinaryKey + ?Sized,
             V: BinaryValue { ... }
    fn get_sparse_list<I, V>(
        &self,
        addr: I,
    ) -> SparseListIndex<Self::Base<'_>, V>
       where I: Into<IndexAddress>,
             V: BinaryValue { ... }
    fn get_key_set<I, K>(&self, addr: I) -> KeySetIndex<Self::Base<'_>, K>
       where I: Into<IndexAddress>,
             K: BinaryKey + ?Sized { ... }
    fn index_type<I>(&self, addr: I) -> Option<IndexType>
       where I: Into<IndexAddress> { ... }
}

Extension trait allowing for easy access to indexes from any type implementing Access.

Implementation details

This trait is essentially a thin wrapper around FromAccess. Where FromAccess returns an access error, the methods of this trait will unwrap() the error and panic.

Required Associated Types

type Base<'a>: RawAccess where Self: 'a

Shortcut for Self::Ref::Base, similar to Item in IntoIterator. Allows to express type boundaries more concisely.

type Ref<'a>: Access<Base = Self::Base<'a>> where Self: 'a

Access that this extension is based on: either Self or &'a Self.

Required Methods

fn get_ref(&self) -> Self::Ref<'_>

Provided Methods

fn get_group<'s, K, I>( &'s self, name: impl Into<String>, ) -> Group<Self::Ref<'s>, K, I>
where K: BinaryKey + ?Sized, I: FromAccess<Self::Ref<'s>>,

Returns a group of indexes. All indexes in the group have the same type. Indexes are initialized lazily; i.e., no initialization is performed when the group is created.

Note that unlike other methods, this one requires address to be a string. This is to prevent collisions among groups.

fn get_entry<I, V>(&self, addr: I) -> Entry<Self::Base<'_>, V>
where I: Into<IndexAddress>, V: BinaryValue,

Gets an entry index with the specified address.

Panics

If the index exists, but is not an entry.

fn get_list<I, V>(&self, addr: I) -> ListIndex<Self::Base<'_>, V>
where I: Into<IndexAddress>, V: BinaryValue,

Gets a list index with the specified address.

Panics

If the index exists, but is not a list.

fn get_map<I, K, V>(&self, addr: I) -> MapIndex<Self::Base<'_>, K, V>
where I: Into<IndexAddress>, K: BinaryKey + ?Sized, V: BinaryValue,

Gets a map index with the specified address.

Panics

If the index exists, but is not a map.

fn get_sparse_list<I, V>(&self, addr: I) -> SparseListIndex<Self::Base<'_>, V>
where I: Into<IndexAddress>, V: BinaryValue,

Gets a sparse list index with the specified address.

Panics

If the index exists, but is not a sparse list.

fn get_key_set<I, K>(&self, addr: I) -> KeySetIndex<Self::Base<'_>, K>
where I: Into<IndexAddress>, K: BinaryKey + ?Sized,

Gets a key set index with the specified address.

Panics

If the index exists, but is not a key set.

fn index_type<I>(&self, addr: I) -> Option<IndexType>
where I: Into<IndexAddress>,

Gets index type at the specified address, or None if there is no index.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.