matterdb::access

Trait FromAccess

pub trait FromAccess<T: Access>: Sized {
    // Required method
    fn from_access(access: T, addr: IndexAddress) -> Result<Self, AccessError>;

    // Provided method
    fn from_root(access: T) -> Result<Self, AccessError> { ... }
}

Constructs an object atop the database. The constructed object provides access to data in the DB, akin to an object-relational mapping.

The access to DB can be readonly or read-write, depending on the T: Access type param. Most object should implement FromAccess<T> for all T: Access.

Simplest FromAccess implementors are indexes; it is also implemented for Group. FromAccess can be implemented for more complex components. Thus, FromAccess can be used to compose storage objects from simpler ones.

Examples

Component with two inner indexes. FromAccess is automatically derived using the matterdb_derive crate.

use matterdb_derive::FromAccess;

#[derive(FromAccess)]
struct InsertOnlyMap<T: Access> {
    map: MapIndex<T::Base, str, String>,
    len: Entry<T::Base, u64>,
}

impl<T: Access> InsertOnlyMap<T>
where
    T::Base: RawAccessMut,
{
    fn insert(&mut self, key: &str, value: String) -> bool {
        if self.map.contains(key) { return false; }
        self.map.put(&key.to_owned(), value);
        self.len.set(self.len.get().unwrap_or_default() + 1);
        true
    }
}

let db = TemporaryDB::new();
let fork = db.fork();
let mut map = InsertOnlyMap::from_access(&fork, "test".into())?;
map.insert("foo", "FOO".to_owned());
map.insert("bar", "BAR".to_owned());
assert_eq!(map.len.get(), Some(2));

// Components could be used with `Group` out of the box:
let group_of_maps: Group<_, u16, InsertOnlyMap<_>> =
    fork.get_group("test_group");
group_of_maps.get(&1).insert("baz", "BAZ".to_owned());
group_of_maps.get(&2).insert("baz", "BUZZ".to_owned());

Required Methods

fn from_access(access: T, addr: IndexAddress) -> Result<Self, AccessError>

Constructs the object at the given address.

Return value

Returns the constructed object.

Errors

An error should be returned if the object cannot be constructed.

Provided Methods

fn from_root(access: T) -> Result<Self, AccessError>

Constructs the object from the root of the access.

The default implementation uses Self::from_access() with an empty address.

Errors

Returns access errors.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.