matterdb

Module indexes

All available MerkleDB indexes.

Structs

Entries

Iterator over key-value pairs of an index.

Entry

An index that may only contain one element.

Group

Group of indexes distinguished by a prefix.

KeySetIndex

A set of key items.

Keys

Iterator over keys of an index.

ListIndex

A list of items where elements are added to the end of the list and are removed starting from the end of the list.

MapIndex

A map of keys and values. Access to the elements of this map is obtained using the keys.

SparseListIndex

A list of items similar to ListIndex; however, it may contain “spaces”. For instance, a list might contain six elements with indexes: 1, 2, 3, 5, 7, 8 (missing 4 and 6). And if you try to get the element for index 4 or 6, you will get None.

Values

Iterator over values of an index.

Traits

IndexIterator

Database object that supports iteration and continuing iteration from an intermediate position.