matterdb::indexes

Struct Entry

pub struct Entry<T: RawAccess, V> { /* private fields */ }

An index that may only contain one element.

You can add an element to this index and check whether it exists. A value should implement BinaryValue trait.

Implementations

impl<T, V> Entry<T, V>
where T: RawAccess, V: BinaryValue,

pub fn get(&self) -> Option<V>

Returns a value of the entry or None if does not exist.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, Entry};

let db = TemporaryDB::new();
let fork = db.fork();
let mut index = fork.get_entry("name");
assert_eq!(None, index.get());

index.set(10);
assert_eq!(Some(10), index.get());

pub fn exists(&self) -> bool

Returns true if a value of the entry exists.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, Entry};

let db = TemporaryDB::new();
let fork = db.fork();
let mut index = fork.get_entry("name");
assert!(!index.exists());

index.set(10);
assert!(index.exists());

impl<T, V> Entry<T, V>
where T: RawAccessMut, V: BinaryValue,

pub fn set(&mut self, value: V)

Changes a value of the entry.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, Entry};

let db = TemporaryDB::new();
let fork = db.fork();
let mut index = fork.get_entry("name");

index.set(10);
assert_eq!(Some(10), index.get());

pub fn remove(&mut self)

Removes a value of the entry.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, Entry};

let db = TemporaryDB::new();
let fork = db.fork();
let mut index = fork.get_entry("name");

index.set(10);
assert_eq!(Some(10), index.get());

index.remove();
assert_eq!(None, index.get());

pub fn take(&mut self) -> Option<V>

Takes the value out of the entry, leaving a None in its place.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, Entry};

let db = TemporaryDB::new();
let fork = db.fork();
let mut index = fork.get_entry("name");

index.set(10);
assert_eq!(Some(10), index.get());

let value = index.take();
assert_eq!(Some(10), value);
assert_eq!(None, index.get());

pub fn swap(&mut self, value: V) -> Option<V>

Replaces the value in the entry with the given one, returning the previously stored value.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, Entry};

let db = TemporaryDB::new();
let fork = db.fork();
let mut index = fork.get_entry("name");

index.set(10);
assert_eq!(Some(10), index.get());

let value = index.swap(20);
assert_eq!(Some(10), value);
assert_eq!(Some(20), index.get());