matterdb::indexes

Struct MapIndex

pub struct MapIndex<T: RawAccess, K: ?Sized, V> { /* private fields */ }

A map of keys and values. Access to the elements of this map is obtained using the keys.

MapIndex requires that keys implement the BinaryKey trait and values implement the BinaryValue trait.

Implementations

impl<T, K, V> MapIndex<T, K, V>
where T: RawAccess, K: BinaryKey + ?Sized, V: BinaryValue,

pub fn get(&self, key: &K) -> Option<V>

Returns a value corresponding to the key.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let mut index = fork.get_map("name");
assert!(index.get(&1).is_none());

index.put(&1, 2);
assert_eq!(Some(2), index.get(&1));

pub fn contains(&self, key: &K) -> bool

Returns true if the map contains a value corresponding to the specified key.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let mut index = fork.get_map("name");
assert!(!index.contains(&1));

index.put(&1, 2);
assert!(index.contains(&1));

pub fn iter(&self) -> Entries<'_, K, V>

Returns an iterator over the entries of the map in ascending order.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let index: MapIndex<_, u8, u8> = fork.get_map("name");

for v in index.iter() {
    println!("{:?}", v);
}

pub fn keys(&self) -> Keys<'_, K>

Returns an iterator over the keys of a map in ascending order.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let index: MapIndex<_, u8, u8> = fork.get_map("name");

for key in index.keys() {
    println!("{}", key);
}

pub fn values(&self) -> Values<'_, V>

Returns an iterator over the values of a map in ascending order of keys.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let index: MapIndex<_, u8, u8> = fork.get_map("name");

for val in index.values() {
    println!("{}", val);
}

pub fn iter_from(&self, from: &K) -> Entries<'_, K, V>

Returns an iterator over the entries of a map in ascending order starting from the specified key.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let index: MapIndex<_, u8, u8> = fork.get_map("name");

for v in index.iter_from(&2) {
    println!("{:?}", v);
}

pub fn keys_from(&self, from: &K) -> Keys<'_, K>

Returns an iterator over the keys of a map in ascending order starting from the specified key.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let index: MapIndex<_, u8, u8> = fork.get_map("name");

for key in index.keys_from(&2) {
    println!("{}", key);
}

pub fn values_from(&self, from: &K) -> Values<'_, V>

Returns an iterator over the values of a map in ascending order of keys starting from the specified key.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let index: MapIndex<_, u8, u8> = fork.get_map("name");
for val in index.values_from(&2) {
    println!("{}", val);
}

impl<T, K, V> MapIndex<T, K, V>
where T: RawAccessMut, K: BinaryKey + ?Sized, V: BinaryValue,

pub fn put(&mut self, key: &K, value: V)

Inserts a key-value pair into a map.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let mut index = fork.get_map("name");

index.put(&1, 2);
assert!(index.contains(&1));

pub fn remove<Q>(&mut self, key: &Q)
where K: Borrow<Q>, Q: BinaryKey + ?Sized,

Removes a key from a map.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let mut index = fork.get_map("name");

index.put(&1, 2);
assert!(index.contains(&1));

index.remove(&1);
assert!(!index.contains(&1));

pub fn clear(&mut self)

Clears a map, removing all entries.

Notes

Currently, this method is not optimized to delete a large set of data. During the execution of this method, the amount of allocated memory is linearly dependent on the number of elements in the index.

Examples

use matterdb::{access::AccessExt, TemporaryDB, Database, MapIndex};

let db = TemporaryDB::default();
let fork = db.fork();
let mut index = fork.get_map("name");

index.put(&1, 2);
assert!(index.contains(&1));

index.clear();
assert!(!index.contains(&1));