matterdb::indexes

Trait IndexIterator

pub trait IndexIterator {
    type Key: BinaryKey + ?Sized;
    type Value: BinaryValue;

    // Required method
    fn index_iter(
        &self,
        from: Option<&Self::Key>,
    ) -> Entries<'_, Self::Key, Self::Value> ;
}

Database object that supports iteration and continuing iteration from an intermediate position.

This trait is implemented for all index collections (i.e., all index types except for Entry) and can thus be used by the generic iteration routines.

Required Associated Types

type Key: BinaryKey + ?Sized

Type encompassing index keys.

type Value: BinaryValue

Type encompassing index values.

Required Methods

fn index_iter( &self, from: Option<&Self::Key>, ) -> Entries<'_, Self::Key, Self::Value>

Continues iteration from the specified position. If from is None, starts the iteration from scratch.