matterdb::migration

Struct Migration

pub struct Migration<T> { /* private fields */ }

Access to migrated indexes.

Migration is conceptually similar to a Prefixed access. For example, an index with address "list" in a migration Migration::new("foo", _) will map to the address "foo.list" after the migration is flushed. The major difference with Prefixed is that the indexes in a migration cannot be accessed in any other way. That is, it is impossible to access an index in a migration without constructing a Migration object first.

Implementations

impl<T: RawAccess> Migration<T>

pub fn new(namespace: impl Into<String>, access: T) -> Self

Creates a migration in the specified namespace.

impl<T: RawAccessMut> Migration<T>

pub fn create_tombstone<I>(&self, addr: I)
where I: Into<IndexAddress>,

Marks an index with the specified address as removed during migration.

Panics

Panics if an index already exists at the specified address.