matterdb

Module rocksdb

An implementation of RocksDB database.

Modules

backup

Backup-related stuff for RocksDB database.

Structs

RocksDB

Database implementation on top of RocksDB backend.

RocksDBSnapshot

A snapshot of a RocksDB.

Constants

ID_SIZE

Size of a byte representation of an index ID, which is used to prefix index keys in a column family.

Functions

next_id_bytes

Generates the sequence of bytes lexicographically following the provided one. Assumes that the provided sequence is less than [u8::MAX; ID_SIZE].