matterdb

Struct Fork

pub struct Fork { /* private fields */ }

A combination of a database snapshot and changes on top of it.

A Fork provides both immutable and mutable operations over the database by implementing the RawAccessMut trait. Like Snapshot, Fork provides read isolation. When mutable operations are applied to a fork, the subsequent reads act as if the changes are applied to the database; in reality, these changes are accumulated in memory.

To apply the changes to the database, you need to convert a Fork into a Patch using Self::into_patch() and then atomically merge it into the database. If two conflicting forks are merged into a database, this can lead to an inconsistent state. If you need to consistently apply several sets of changes to the same data, the next fork should be created after the previous fork has been merged.

Fork also supports checkpoints (flush and rollback methods), which allows rolling back the latest changes. A checkpoint is created automatically after calling the flush method.

let db = TemporaryDB::new();
let mut fork = db.fork();
fork.get_list("list").extend(vec![1_u32, 2]);
fork.flush();
fork.get_list("list").push(3_u32);
fork.rollback();
// The changes after the latest `flush()` are now forgotten.
let list = fork.get_list::<_, u32>("list");
assert_eq!(list.len(), 2);

In order to convert a fork into &dyn Snapshot presentation, convert it into a Patch and use a reference to it (Patch implements Snapshot). Using <Fork as RawAccess>::snapshot for this purpose is logically incorrect and may lead to hard-to-debug errors.

Borrow checking

It is possible to create only one instance of index with the specified IndexAddress based on a single fork. If an additional instance is requested, the code will panic in runtime. Hence, obtaining indexes from a Fork functions similarly to RefCell::borrow_mut().

For example the code below will panic at runtime.

let db = TemporaryDB::new();
let fork = db.fork();
let index = fork.get_list::<_, u8>("index");
// This code will panic at runtime.
let index2 = fork.get_list::<_, u8>("index");

To enable immutable / shared references to indexes, you may use readonly() method:

let db = TemporaryDB::new();
let fork = db.fork();
fork.get_list::<_, u8>("index").extend(vec![1, 2, 3]);

let readonly = fork.readonly();
let index = readonly.get_list::<_, u8>("index");
// Works fine.
let index2 = readonly.get_list::<_, u8>("index");

It is impossible to mutate index contents having a readonly access to the fork; this is checked by the Rust type system.

Shared references work like RefCell::borrow(); it is a runtime error to try to obtain a shared reference to an index if there is an exclusive reference to the same index, and vice versa.

Implementations

impl Fork

pub fn flush(&mut self)

Finalizes all changes that were made after previous execution of the flush method. If no flush method had been called before, finalizes all changes that were made after creation of Fork.

pub fn rollback(&mut self)

Rolls back all changes that were made after the latest execution of the flush method.

pub fn into_patch(self) -> Patch

Converts the fork into Patch consuming the fork instance.

pub fn readonly(&self) -> ReadonlyFork<'_>

Returns a readonly wrapper around the fork. Indexes created based on the readonly version cannot be modified; on the other hand, it is possible to have multiple copies of an index at the same time.