matterdb

Struct IndexAddress

pub struct IndexAddress { /* private fields */ }

Represents the address of an index in the database.

An address has a string name and an optional byte key. An index is uniquely identified by its address. Different addresses correspond to different indexes. Addresses with the same name and differing keys are said to belong to the same group (see also Group). Groups can be used for a potentially unbounded group of indexes that can be identified by a certain key (for example, ListIndex with the transaction history of a wallet keyed by the PublicKey of the wallet).

In contrast with ResolvedAddress, IndexAddress is a high-level logical construct; it does not directly map to key-value storage abstractions (column families and their keys).

Examples

IndexAddress can be used implicitly, since &str and (&str, &impl BinaryKey) can both be converted into an address.

use matterdb::{access::AccessExt, IndexAddress, TemporaryDB, Database};

let db = TemporaryDB::new();
let fork = db.fork();

// Using a string address:
let map = fork.get_map::<_, String, u8>("map");
// Using an address within an index family:
let list = fork.get_list::<_, String>(("index", &3_u32));
// Using `IndexAddress` explicitly:
let addr = IndexAddress::from_root("data").append_key(&vec![1, 2, 3]);
let set = fork.get_key_set::<_, u64>(addr);

Implementations

impl IndexAddress

pub fn from_root<S: Into<String>>(root: S) -> Self

Creates new IndexAddress with the specified name.

pub fn name(&self) -> &str

Returns the name part of IndexAddress.

pub fn id_in_group(&self) -> Option<&[u8]>

Returns the bytes part of IndexAddress.

pub fn prepend_name(self, prefix: &str) -> Self

Prepends a name part to IndexAddress. The name is separated from the existing name by a dot ..

Examples

let addr = IndexAddress::from_root("foo");
let prefixed = addr.prepend_name("prefix");
assert_eq!(prefixed.name(), "prefix.foo");

pub fn append_name(self, suffix: &str) -> Self

Appends a name part to IndexAddress. The name is separated from the existing name by a dot ..

Examples

let addr = IndexAddress::from_root("foo");
let suffixed = addr.append_name("suffix");
assert_eq!(suffixed.name(), "foo.suffix");

pub fn append_key<K: BinaryKey + ?Sized>(self, suffix: &K) -> Self

Appends a key to the IndexAddress.