matterdb

Trait BinaryKey

pub trait BinaryKey: ToOwned {
    // Required methods
    fn size(&self) -> usize;
    fn write(&self, buffer: &mut [u8]) -> usize;
    fn read(buffer: &[u8]) -> Self::Owned;
}

A type that can be (de)serialized as a key in the blockchain storage.

Since keys are sorted in the serialized form, the big-endian encoding should be used with unsigned integer types. Note, however, that the big-endian encoding will not sort signed integer types in the natural order; therefore, they are mapped to the corresponding unsigned type by adding a constant to the source value.

Examples

use std::mem;
use matterdb::BinaryKey;

#[derive(Clone)]
struct Key {
    a: i16,
    b: u32,
}

impl BinaryKey for Key {
    fn size(&self) -> usize {
        mem::size_of_val(&self.a) + mem::size_of_val(&self.b)
    }

    fn write(&self, buffer: &mut [u8]) -> usize {
        self.a.write(&mut buffer[0..2]);
        self.b.write(&mut buffer[2..6]);
        self.size()
    }

    fn read(buffer: &[u8]) -> Self {
        let a = i16::read(&buffer[0..2]);
        let b = u32::read(&buffer[2..6]);
        Key { a, b }
    }
}

Required Methods

fn size(&self) -> usize

Returns the size of the serialized key in bytes.

fn write(&self, buffer: &mut [u8]) -> usize

Serializes the key into the specified buffer of bytes.

The caller must guarantee that the size of the buffer is equal to the precalculated size of the serialized key returned via size(). Returns number of written bytes. The provided buffer may be uninitialized; an implementor must not read from it.

fn read(buffer: &[u8]) -> Self::Owned

Deserializes the key from the specified buffer of bytes.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.