matterdb

Trait BinaryValue

pub trait BinaryValue: Sized {
    // Required methods
    fn to_bytes(&self) -> Vec<u8> ;
    fn from_bytes(bytes: Cow<'_, [u8]>) -> Result<Self>;

    // Provided method
    fn into_bytes(self) -> Vec<u8>  { ... }
}

A type that can be (de)serialized as a value in the blockchain storage.

If you need to implement BinaryValue for your types, use little-endian encoding for integer types for compatibility with modern architectures.

Examples

Implementing BinaryValue for the type:

use std::borrow::Cow;
use matterdb::BinaryValue;

#[derive(Clone)]
struct Data {
    a: i16,
    b: u32,
}

impl BinaryValue for Data {
    fn to_bytes(&self) -> Vec<u8> {
        let mut buf = vec![0_u8; 6];
        buf[0..2].copy_from_slice(&self.a.to_le_bytes());
        buf[2..6].copy_from_slice(&self.b.to_le_bytes());
        buf
    }

    fn from_bytes(bytes: Cow<[u8]>) -> anyhow::Result<Self> {
        let mut buf = bytes.as_ref();
        let a = i16::from_le_bytes(buf[0..2].try_into().unwrap());
        let b = u32::from_le_bytes(buf[2..6].try_into().unwrap());
        Ok(Self { a, b })
    }
}

Required Methods

fn to_bytes(&self) -> Vec<u8>

Serializes the given value to the vector of bytes.

fn from_bytes(bytes: Cow<'_, [u8]>) -> Result<Self>

Deserializes the value from the given bytes array.

Errors

Should return an error if bytes do not represent a valid value.

Provided Methods

fn into_bytes(self) -> Vec<u8>

Consumes and serializes the given value to the vector of bytes. This method is faster with the wrapped values, thus if you wouldn’t use value after serialization use it.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.